TURMERIC-1576: Ignore report generation when src/main/java isn't around.
With the recent changes to the cobertura profile, it tries to specify where
the source code is.  If there is no src/main/java defined for the project
it will cause the antrun task to die during report generation.  To correct
this we use the antcontrib try/catch to catch the error and allow processing
to continue.  Thanks to the following tip on StackOverflow.

http://stackoverflow.com/questions/2666794/how-do-i-prevent-maven-build-fail-on-ant-task-failure