TURMERIC-1665: update parent pom to not regenerate sources during package phase
Currently the jar and test-jar goals are used with the maven-source-plugin
however these goals cause the generate-sources phase to run again during
the build. This is causing an issue with codegen when the jvm is reforked
and getting it confused on what Xerces implementation to use.  To solve this
we need to use the jar-no-fork and test-jar-no-fork.  These accomplish the
same thing, but do not run the generate-sources Lifecycle again.